feat: Add analysis type functionality
